+++ crates/macros/src/config_enum/variant.rs
@@ -1,4 +1,4 @@
-use crate::utils::{extract_common_attrs, format_case};
+use crate::utils::{extract_comment, extract_common_attrs, format_case, has_deprecated_attr};
 use darling::FromAttributes;
 use proc_macro2::{Ident, TokenStream};
 use quote::quote;
@@ -106,16 +106,52 @@ impl<'l> Variant<'l> {
         }
     }
 
+    pub fn get_schema_type(&self) -> TokenStream {
+        let name = self.name.to_string();
+        let value = &self.value;
+        let deprecated = has_deprecated_attr(&self.attrs);
+
+        let description = if let Some(comment) = extract_comment(&self.attrs) {
+            quote! {
+                Some(#comment.into())
+            }
+        } else {
+            quote! {
+                None
+            }
+        };
+
+        let type_of = if self.args.fallback {
+            quote! {
+                SchemaType::string()
+            }
+        } else {
+            quote! {
+                SchemaType::literal(LiteralValue::String(#value.into()))
+            }
+        };
+
+        quote! {
+            SchemaField {
+                name: Some(#name.into()),
+                description: #description,
+                type_of: #type_of,
+                deprecated: #deprecated,
+                ..Default::default()
+            }
+        }
+    }
+
     pub fn get_unit_name(&self) -> TokenStream {
         let name = &self.name;
 
         if self.args.fallback {
             quote! {
-                Self::#name(Default::default()),
+                Self::#name(Default::default())
             }
         } else {
             quote! {
-                Self::#name,
+                Self::#name
             }
         }
     }