+++ crates/types/src/structs.rs
@@ -1,0 +1,9 @@
+use crate::SchemaField;
+
+#[derive(Clone, Debug, Default, Eq, PartialEq)]
+pub struct StructType {
+    pub additional: bool,
+    pub fields: Vec<SchemaField>,
+    pub name: String,
+    pub required: Vec<String>,
+}